// services/interfaces/Professors.ts
export interface Professor {
    id: number;
    name: string;
    middleName: string;
    lastName: string;
    email: string;
    gender: string;
    rfc: string;
    curp: string;
    status: string;
    department: string;
}

export interface ProfessorForm {
    name: string;
    middleName: string;
    lastName: string;
    email: string;
    gender: string;
    rfc: string;
    curp: string;
    status: string;
    department: string;
}